refactor: repackage 1. shared아래 vo가 없어서 수정
