+++ application-module/src/main/kotlin/com/reservation/user/history/access/port/input/CreateUserAccessHistoriesCommand.kt
@@ -2,7 +2,7 @@ package com.reservation.user.history.access.port.input
 
 import com.reservation.enumeration.AccessStatus
 import com.reservation.user.history.access.port.output.CreateUserAccessHistories.CreateUserHistoryInquiry
-import com.reservation.user.shared.LoginId
+import com.reservation.user.shared.vo.LoginId
 import java.time.LocalDateTime
 
 fun interface CreateUserAccessHistoriesCommand {
