+++ application-module/src/main/kotlin/com/reservation/user/history/access/port/output/CreateUserAccessHistories.kt
@@ -1,7 +1,7 @@
 package com.reservation.user.history.access.port.output
 
 import com.reservation.enumeration.AccessStatus
-import com.reservation.user.shared.LoginId
+import com.reservation.user.shared.vo.LoginId
 import java.time.LocalDateTime
 
 fun interface CreateUserAccessHistories {
