+++ application-module/src/main/kotlin/com/reservation/user/resign/port/output/ResignTargetUser.kt
@@ -1,7 +1,7 @@
 package com.reservation.user.resign.port.output
 
 import com.reservation.user.resign.EncryptedAttributes
-import com.reservation.user.shared.LoginId
+import com.reservation.user.shared.vo.LoginId
 import java.time.LocalDateTime
 
 fun interface ResignTargetUser {
