+++ application-module/src/test/kotlin/com/reservation/user/self/usecase/ChangeGeneralUserNicknameUseCaseTest.kt
@@ -14,9 +14,9 @@ import com.reservation.user.self.port.output.CheckGeneralUserNicknameDuplicated
 import com.reservation.user.self.port.output.LoadGeneralUser
 import com.reservation.user.self.port.output.LoadGeneralUser.LoadGeneralUserResult
 import com.reservation.user.service.ChangeUserNicknameService
-import com.reservation.user.shared.LoginId
-import com.reservation.user.shared.Password
-import com.reservation.user.shared.PersonalAttributes
+import com.reservation.user.shared.vo.LoginId
+import com.reservation.user.shared.vo.Password
+import com.reservation.user.shared.vo.PersonalAttributes
 import io.mockk.Runs
 import io.mockk.every
 import io.mockk.impl.annotations.InjectMockKs
