+++ core-module/src/main/kotlin/com/reservation/user/resign/ResignedUser.kt
@@ -1,6 +1,6 @@
 package com.reservation.user.resign
 
-import com.reservation.user.shared.LoginId
+import com.reservation.user.shared.vo.LoginId
 import java.time.LocalDateTime
 
 class ResignedUser(
