+++ core-module/src/main/kotlin/com/reservation/user/shared/vo/PersonalAttributes.kt
@@ -1,4 +1,4 @@
-package com.reservation.user.shared
+package com.reservation.user.shared.vo
 
 data class PersonalAttributes(
     val email: String,
