+++ core-module/src/test/kotlin/com/reservation/authenticate/SignInSuccessTest.kt
@@ -7,8 +7,8 @@ import com.reservation.authenticate.service.AuthenticateSignInService
 import com.reservation.enumeration.AccessStatus
 import com.reservation.enumeration.UserStatus.ACTIVATED
 import com.reservation.fixture.FixtureMonkeyFactory
-import com.reservation.user.shared.LockState
-import com.reservation.user.shared.Password
+import com.reservation.user.shared.vo.LockState
+import com.reservation.user.shared.vo.Password
 import com.reservation.utilities.encrypt.password.PasswordEncoderUtility
 import io.kotest.core.spec.style.BehaviorSpec
 import io.kotest.matchers.collections.shouldHaveSize
