+++ core-module/src/test/kotlin/com/reservation/user/general/ChangeGeneralUserNicknameServiceFailureTest.kt
@@ -6,9 +6,9 @@ import com.reservation.fixture.FixtureMonkeyFactory
 import com.reservation.user.common.exceptions.InvalidateUserElementException
 import com.reservation.user.self.User
 import com.reservation.user.service.ChangeUserNicknameService
-import com.reservation.user.shared.LoginId
-import com.reservation.user.shared.Password
-import com.reservation.user.shared.PersonalAttributes
+import com.reservation.user.shared.vo.LoginId
+import com.reservation.user.shared.vo.Password
+import com.reservation.user.shared.vo.PersonalAttributes
 import io.kotest.assertions.throwables.shouldThrow
 import io.kotest.core.spec.style.FunSpec
