+++ core-module/src/test/kotlin/com/reservation/user/general/ChangeGeneralUserNicknameServiceTest.kt
@@ -5,9 +5,9 @@ import com.reservation.fixture.CommonlyUsedArbitraries
 import com.reservation.fixture.FixtureMonkeyFactory
 import com.reservation.user.self.User
 import com.reservation.user.service.ChangeUserNicknameService
-import com.reservation.user.shared.LoginId
-import com.reservation.user.shared.Password
-import com.reservation.user.shared.PersonalAttributes
+import com.reservation.user.shared.vo.LoginId
+import com.reservation.user.shared.vo.Password
+import com.reservation.user.shared.vo.PersonalAttributes
 import io.kotest.core.spec.style.BehaviorSpec
 import io.kotest.matchers.shouldBe
