+++ core-module/src/test/kotlin/com/reservation/user/resign/ResignUserServiceTest.kt
@@ -3,9 +3,9 @@ package com.reservation.user.resign
 import com.reservation.fixture.CommonlyUsedArbitraries
 import com.reservation.user.resign.service.ResignUserService
 import com.reservation.user.self.User
-import com.reservation.user.shared.LoginId
-import com.reservation.user.shared.Password
-import com.reservation.user.shared.PersonalAttributes
+import com.reservation.user.shared.vo.LoginId
+import com.reservation.user.shared.vo.Password
+import com.reservation.user.shared.vo.PersonalAttributes
 import com.reservation.utilities.encrypt.bidirectional.BidirectionalEncryptUtility
 import com.reservation.utilities.generator.uuid.UuidGenerator
 import io.kotest.core.spec.style.BehaviorSpec
